// src/enum/awsOfferCloudDetails.enum.ts
export enum awsOfferCloudDetails {
  DEAL_NAME = "Name",
  DESCRIPTION = "Description",
  SALES_NAME = "Primary Sales Agent Name",
  SALES_EMAIL = "Primary Sales Agent Email",
  OFFER_ID = "Offer ID",
  OFFER_TYPE = "Offer Type",
  OFFER_URL = "Offer URL",
  STAGE_NAME = "Stage Name",
  CUSTOMER_COMPANY = "Customer Company Name",
  CUSTOMER_EMAIL = "Customer Email",
  ADDRESS = "Address",
  COUNTRY = "Country",
  STATE = "State",
  ZIP_CODE = "Zip/Postal Code",
  AWS_ID = "AWS Account ID",
  PRODUCT = "Product",
  PRICING_MODEL = "Product Pricing Model",
  CONTRACT_MODEL = "Contract model",
  USAGE_MODEL = "Usage model",
  PRODUCT_ID = "Product ID",
  FPS = "Flexible Payment Schedule",
  OFFER_CREATED = "Offer Created Date",
  OFFER_EXPERIED = "Offer Expiration Date",
  RENEWAL_STATUS = "Renewal Status",
  RENEWAL_OFFER = "Renewal offer",
  NON_RENEWAL_OFFER = "Non-Renewal offer",
  RENEWAL_TYPE = "Renewal Type",
  CONTRACT_END_DATE = "Contract End Date",
  CONTRACT_START_DATE = "Contract Start Date",
  USAGE_START_DATE = "Usage Start Date",
  AT_ACCEPTANCE = "At Acceptance",
  CONTRACT_DURATION = "Contract Duration",
  USAGE_DURATION = "Usage Duration",
  USAGE_END_DATE = "Usage End Date",
  LEGAL_TERM_TYPE = "Legal Term Type",
  TERMS_CONDITIONS = "Terms and Conditions",
  PURCHASING_OPTIONS = "Purchasing options",
  MULTIPLE_DIMENSIONS = "Multiple dimension per contract",
  SINGLE_DIMENSIONS = "Single dimension per contract",
}